Reduced Equipment Build Costs
It is by the sacred decree of the Holy Polarace and his council of tech weirdos; Krco, Lavrentis, and Star that all weapon fabrication costs shall be cut in half.
This edict shall now and forever ensure that armies shall be larger and the 3rd global crusade shall be ever more grand.
The Order of Tech Weirdos has spoken!